Modularisasi JavaScript files
